that was a lot of stuff
