+++ crates/sui-indexer-alt-e2e-tests/Cargo.toml
@@ -25,7 +25,7 @@ url.workspace = true
 
 simulacrum.workspace = true
 sui-indexer-alt.workspace = true
-sui-indexer-alt-framework.workspace = true
+sui-indexer-alt-framework = { workspace = true, features = ["postgres"] }
 sui-indexer-alt-jsonrpc.workspace = true
 sui-pg-db.workspace = true
 sui-types.workspace = true
