+++ crates/sui-indexer-alt/Cargo.toml
@@ -32,7 +32,7 @@ tracing.workspace = true
 url.workspace = true
 
 sui-default-config.workspace = true
-sui-indexer-alt-framework.workspace = true
+sui-indexer-alt-framework = { workspace = true, features = ["postgres"] }
 sui-indexer-alt-metrics.workspace = true
 sui-indexer-alt-schema.workspace = true
 sui-protocol-config.workspace = true
