+++ crates/sui-indexer-alt/src/bootstrap.rs
@@ -6,8 +6,8 @@ use std::time::Duration;
 use anyhow::{bail, Context, Result};
 use diesel::{OptionalExtension, QueryDsl, SelectableHelper};
 use diesel_async::RunQueryDsl;
+use sui_indexer_alt_framework::postgres::PgStore;
 use sui_indexer_alt_framework::store::Store;
-use sui_indexer_alt_framework::sui_indexer_alt_framework_store_pg::pg_store::PgStore;
 use sui_indexer_alt_framework::types::{
     full_checkpoint_content::CheckpointData,
     sui_system_state::{get_sui_system_state, SuiSystemStateTrait},
