+++ crates/sui-indexer-alt/src/handlers/tx_calls.rs
@@ -9,8 +9,8 @@ use diesel::{ExpressionMethods, QueryDsl};
 use diesel_async::RunQueryDsl;
 use sui_indexer_alt_framework::{
     pipeline::{concurrent::Handler, Processor},
+    postgres::PgStore,
     store::Store,
-    sui_indexer_alt_framework_store_pg::pg_store::PgStore,
     types::{full_checkpoint_content::CheckpointData, transaction::TransactionDataAPI},
 };
 use sui_indexer_alt_schema::{schema::tx_calls, transactions::StoredTxCalls};
@@ -97,7 +97,7 @@ mod tests {
     use diesel_async::RunQueryDsl;
     use sui_indexer_alt_framework::{
         types::{base_types::ObjectID, test_checkpoint_data_builder::TestCheckpointDataBuilder},
-        Indexer,
+        Indexer, IndexerPostgresExt,
     };
     use sui_indexer_alt_schema::MIGRATIONS;
