+++ crates/sui-indexer-alt/src/main.rs
@@ -13,9 +13,9 @@ use sui_indexer_alt::args::Command;
 use sui_indexer_alt::config::IndexerConfig;
 use sui_indexer_alt::config::Merge;
 use sui_indexer_alt::setup_indexer;
-use sui_indexer_alt_framework::db::reset_database;
-use sui_indexer_alt_framework::sui_indexer_alt_framework_store_pg::pg_store::PgStore;
+use sui_indexer_alt_framework::postgres::{reset_database, PgStore};
 use sui_indexer_alt_framework::Indexer;
+use sui_indexer_alt_framework::IndexerPostgresExt;
 use sui_indexer_alt_metrics::MetricsService;
 use sui_indexer_alt_schema::MIGRATIONS;
 use tokio::fs;
